+++ Multiplatform/macOS/Preferences/Preferences Subviews/AdvancedPreferencesView.swift
